% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dat.R
\docType{data}
\name{example_dat}
\alias{example_dat}
\title{A List of Simulated Longitudinal Data}
\format{
This data frame contains the following columns:

\describe{
\item{id}{identifer of child}
\item{time}{an indicator of age intervals (1 = age 2--7 for childhood
BMI z-scores and age 12--17 for obesity in adolescence,
2 = age 7--12 for preadolescence
BMI z-scores and age 20--29 for obesity in adulthood)}
\item{gender}{an indicator of gender (1 = male, 0 = female)}
\item{age}{child's age}
\item{bmi_z}{BMI z-score}
\item{obes}{an indicator of obesity (1 = obese, 0 = none)}
}
}
\usage{
example_dat
}
\description{
The \code{example_dat} is a list of two simulated longitudinal data.
}
\details{
The first object is the \code{biomk} data which mimics
the scaled cerebrospinal fluid biomarker level distribution,
containing the baseline age, gender, biomarker levels, and visit times
for 100 subjects.
The second list object is the \code{primary} data which mimics
the distribution of the cognitive scores,
containing the baseline age, baseline education, gender, cognitive scores,
time for each available score, and a categorical variable indicating if an
individual has more than a copy of a certain gene for the 100 subjects.
}
\examples{
data(example_dat)

}
\keyword{datasets}
