% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlong.R
\name{jlong}
\alias{jlong}
\title{Fit Linear Mixed-Effects Model with Covariate from another Linear Mixed-Effects Model}
\usage{
jlong(
  formula.primod,
  formula.submod,
  data.primod,
  data.submod,
  var.shared,
  ...
)
}
\arguments{
\item{formula.primod}{A two-sided formula for the primary response variable}

\item{formula.submod}{A two-sided formula for the sub model (biomarker data model)}

\item{data.primod}{Primary response data}

\item{data.submod}{Sub model data}

\item{var.shared}{The name of the variable that both models share but remember
to put this name in the formula for the response variable}

\item{...}{}
}
\value{
An object of class \code{jlong}
}
\description{
This function connects two linear mixed-effects models via a joint model.
}
